fix ploting bugs

+++ pi_object.c
@@ -227,7 +227,7 @@ ObjSound *new_sound(Mix_Chunk *chunk)
     return sound;
 }
 
-ObjSprite *new_sprite(uint8_t width, uint8_t height, uint8_t *data)
+ObjSprite *new_sprite(uint16_t width, uint16_t height, uint8_t *data)
 {
     ObjSprite *sprite = CREATE_OBJ(ObjSprite, OBJ_SPRITE);
     sprite->width = width;
